Reorganize query parameters
