+++ app-backend/api/src/main/scala/aoi/Routes.scala
@@ -3,20 +3,25 @@ package com.azavea.rf.api.aoi
 import java.util.UUID
 
 import akka.http.scaladsl.server.Route
-import com.azavea.rf.api.utils.queryparams.QueryParametersCommon
-import com.azavea.rf.common.{Authentication, CommonHandlers, UserErrorHandler}
-import com.azavea.rf.database.tables.AOIs
-import com.azavea.rf.datamodel._
 import com.lonelyplanet.akka.http.extensions.PaginationDirectives
 import de.heikoseeberger.akkahttpcirce.ErrorAccumulatingCirceSupport._
 
-// --- //
+import com.azavea.rf.api.DaoQueryExtension._
+import com.azavea.rf.api.utils.queryparams.QueryParametersCommon
+import com.azavea.rf.common.{Authentication, CommonHandlers, UserErrorHandler}
+import com.azavea.rf.database.tables.AOIs
+import com.azavea.rf.database.AoiDao
+import com.azavea.rf.datamodel._
+import doobie._
+import cats.effect.IO
+
 
 trait AoiRoutes extends Authentication
     with UserErrorHandler
     with QueryParametersCommon
     with PaginationDirectives
     with CommonHandlers {
+  implicit val xa: Transactor[IO] = ???
 
   val aoiRoutes: Route = handleExceptions(userExceptionHandler) {
     pathEndOrSingleSlash {
@@ -34,7 +39,7 @@ trait AoiRoutes extends Authentication
   def listAOIs: Route = authenticate { user =>
     (withPagination & aoiQueryParameters) { (page, aoiQueryParams) =>
       complete {
-        AOIs.listAOIs(page, aoiQueryParams, user)
+        AoiDao.page(aoiQueryParams, user, page)
       }
     }
   }
