+++ app-backend/api/src/main/scala/categories/QueryParameters.scala
@@ -3,7 +3,7 @@ package com.azavea.rf.api.toolcategory
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.directives.ParameterDirectives.parameters
 
-import com.azavea.rf.database.query._
+import com.azavea.rf.datamodel._
 import com.azavea.rf.api.utils.queryparams._
 
 /* Trait to abstract out query parameters for tool categories */
