+++ app-backend/api/src/main/scala/grid/QueryParameters.scala
@@ -5,7 +5,7 @@ import java.util.UUID
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.directives.ParameterDirectives.parameters
 
-import com.azavea.rf.database.query._
+import com.azavea.rf.datamodel._
 import com.azavea.rf.api.image.ImageQueryParametersDirective
 import com.azavea.rf.api.utils.queryparams._
