+++ app-backend/api/src/main/scala/image/QueryParams.scala
@@ -4,9 +4,10 @@ import java.util.UUID
 
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.directives.ParameterDirectives.parameters
-import com.azavea.rf.database.query.{CombinedImageQueryParams, ImageQueryParameters}
 
+import com.azavea.rf.datamodel._
 import com.azavea.rf.api.utils.queryparams._
+
 
 /** Trait to mix in for image specific query parameters */
 trait ImageQueryParametersDirective extends QueryParametersCommon {
