+++ app-backend/api/src/main/scala/maptoken/QueryParams.scala
@@ -4,7 +4,7 @@ import java.util.UUID
 
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.directives.ParameterDirectives.parameters
-import com.azavea.rf.database.query.{CombinedMapTokenQueryParameters, MapTokenQueryParameters}
+import com.azavea.rf.datamodel._
 import com.azavea.rf.api.utils.queryparams._
 
 /** Trait to mix in for image specific query parameters */
