+++ app-backend/api/src/main/scala/thumbnail/QueryParameters.scala
@@ -7,7 +7,7 @@ import java.time.Instant
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.{Directive1, Rejection}
 import akka.http.scaladsl.server.directives.ParameterDirectives.parameters
-import com.azavea.rf.database.query.ThumbnailQueryParameters
+import com.azavea.rf.datamodel._
 
 import com.azavea.rf.api.utils.queryparams._
