+++ app-backend/api/src/main/scala/toolrun/QueryParameters.scala
@@ -4,10 +4,8 @@ import java.util.UUID
 
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.directives.ParameterDirectives.parameters
-import com.azavea.rf.database.query.{CombinedImageQueryParams, ImageQueryParameters}
-
+import com.azavea.rf.datamodel._
 import com.azavea.rf.api.utils.queryparams._
-import com.azavea.rf.database.query.{ToolRunQueryParameters, CombinedToolRunQueryParameters}
 
 trait ToolRunQueryParametersDirective extends QueryParametersCommon {
   val toolRunSpecificQueryParams = parameters((
