+++ app-backend/database/src/main/scala/com/azavea/rf/database/tables/Annotations.scala
@@ -6,7 +6,6 @@ import java.util.UUID
 import com.azavea.rf.database.{Database => DB}
 import com.azavea.rf.database.ExtendedPostgresDriver.api._
 import com.azavea.rf.database.fields.{OrganizationFkFields, TimestampFields, UserFkFields}
-import com.azavea.rf.database.query.{AnnotationQueryParameters, ListQueryResult}
 import com.azavea.rf.datamodel._
 
 import geotrellis.slick.Projected
