+++ app-backend/database/src/main/scala/com/azavea/rf/database/tables/Projects.scala
@@ -10,7 +10,6 @@ import akka.http.scaladsl.model.{IllegalRequestException, StatusCodes}
 import com.azavea.rf.database.{Database => DB}
 import com.azavea.rf.database.ExtendedPostgresDriver.api._
 import com.azavea.rf.database.fields._
-import com.azavea.rf.database.query._
 import com.azavea.rf.database.sort._
 import com.azavea.rf.datamodel._
 import com.azavea.rf.datamodel.color._
