+++ app-backend/database/src/main/scala/com/azavea/rf/database/tables/Tools.scala
@@ -4,7 +4,6 @@ import io.circe._
 import com.azavea.rf.database.fields._
 import com.azavea.rf.database.sort._
 import com.azavea.rf.datamodel._
-import com.azavea.rf.database.query._
 import com.azavea.rf.database.{Database => DB}
 import com.azavea.rf.database.ExtendedPostgresDriver.api._
