+++ app-backend/datamodel/src/main/scala/com/azavea/rf/datamodel/QueryResults.scala
@@ -1,4 +1,4 @@
-package com.azavea.rf.database.query
+package com.azavea.rf.datamodel
 
 import slick.dbio.DBIO
